+++ src/main/java/com/hotels/beeju/ThriftHiveMetaStoreJUnitRule.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2015-2020 Expedia, Inc.
+ * Copyright (C) 2015-2021 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,6 +16,8 @@
 package com.hotels.beeju;
 
 import java.util.Map;
+
+import org.junit.runner.Description;
 
 import com.hotels.beeju.core.ThriftHiveMetaStoreCore;
 
@@ -50,33 +52,45 @@ public class ThriftHiveMetaStoreJUnitRule extends HiveMetaStoreJUnitRule {
    * Create a Thrift Hive Metastore service with a pre-created database using the provided name and configuration.
    *
    * @param databaseName Database name.
-   * @param preConfiguration Hive configuration properties that will be set prior to BeeJU potentially overriding these with its defaults.
+   * @param preConfiguration Hive configuration properties that will be set prior to BeeJU potentially overriding these
+   *          with its defaults.
    */
   public ThriftHiveMetaStoreJUnitRule(String databaseName, Map<String, String> preConfiguration) {
     super(databaseName, preConfiguration);
   }
-  
+
   /**
    * Create a Thrift Hive Metastore service with a pre-created database using the provided name and configuration.
    *
    * @param databaseName Database name.
-   * @param preConfiguration Hive configuration properties that will be set prior to BeeJU potentially overriding these with its defaults.
+   * @param preConfiguration Hive configuration properties that will be set prior to BeeJU potentially overriding these
+   *          with its defaults.
    * @param postConfiguration Hive configuration properties that will be set to override BeeJU's defaults.
    */
-  public ThriftHiveMetaStoreJUnitRule(String databaseName, Map<String, String> preConfiguration, Map<String, String> postConfiguration) {
+  public ThriftHiveMetaStoreJUnitRule(
+      String databaseName,
+      Map<String, String> preConfiguration,
+      Map<String, String> postConfiguration) {
     super(databaseName, preConfiguration, postConfiguration);
   }
 
   @Override
-  protected void before() throws Throwable {
-    thriftHiveMetaStoreCore.initialise();
-    super.before();
+  public void starting(Description description) {
+    try {
+      thriftHiveMetaStoreCore.initialise();
+    } catch (Exception e) {
+      throw new RuntimeException(e);
+    }
+    super.starting(description);
   }
 
   @Override
-  protected void after() {
-    thriftHiveMetaStoreCore.shutdown();
-    super.after();
+  public void finished(Description description) {
+    try {
+      thriftHiveMetaStoreCore.shutdown();
+    } finally {
+      super.finished(description);
+    }
   }
 
   /**
@@ -92,9 +106,10 @@ public class ThriftHiveMetaStoreJUnitRule extends HiveMetaStoreJUnitRule {
   public int getThriftPort() {
     return thriftHiveMetaStoreCore.getThriftPort();
   }
-  
+
   /**
-   * @param thriftPort The Port to use for the Thrift Hive metastore, if not set then a port number will automatically be allocated.
+   * @param thriftPort The Port to use for the Thrift Hive metastore, if not set then a port number will automatically
+   *          be allocated.
    */
   public void setThriftPort(int thriftPort) {
     thriftHiveMetaStoreCore.setThriftPort(thriftPort);