+++ src/test/java/com/hotels/beeju/HiveMetaStoreJUnitRuleTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2015-2019 Expedia, Inc.
+ * Copyright (C) 2015-2021 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,7 +21,7 @@ import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 
 import java.io.File;
-import java.util.HashMap;
+import java.util.Collections;
 import java.util.Map;
 
 import org.apache.hadoop.hive.conf.HiveConf;
@@ -37,61 +37,62 @@ import org.junit.Test;
 public class HiveMetaStoreJUnitRuleTest {
 
   private static File tempRoot;
-  public @Rule HiveMetaStoreJUnitRule hiveDefaultName = new HiveMetaStoreJUnitRule();
-  public @Rule HiveMetaStoreJUnitRule hiveCustomName = new HiveMetaStoreJUnitRule("my_test_database");
+  public @Rule HiveMetaStoreJUnitRule defaultDbRule = new HiveMetaStoreJUnitRule();
+  public @Rule HiveMetaStoreJUnitRule customDbRule = new HiveMetaStoreJUnitRule("my_test_database");
+  public @Rule HiveMetaStoreJUnitRule customPropertiesRule = new HiveMetaStoreJUnitRule("custom_props_database", customConfProperties());
 
+  private Map<String, String> customConfProperties() {
+    return Collections.singletonMap("my.custom.key", "my.custom.value");
+  }
 
   @Before
   public void before() {
-    tempRoot = hiveDefaultName.tempDir();
+    tempRoot = defaultDbRule.tempDir();
     assertTrue(tempRoot.exists());
   }
 
   @Test
   public void hiveDefaultName() throws Exception {
-    assertRuleInitialised(hiveDefaultName);
+    assertRuleInitialised(defaultDbRule);
   }
 
   @Test
   public void hiveCustomName() throws Exception {
-    assertRuleInitialised(hiveCustomName);
+    assertRuleInitialised(customDbRule);
   }
 
   private static void assertRuleInitialised(HiveMetaStoreJUnitRule hive) throws Exception {
     String databaseName = hive.databaseName();
-
     Database database = hive.client().getDatabase(databaseName);
 
     assertThat(database.getName(), is(databaseName));
-    File databaseFolder = new File(hive.tempDir(), databaseName);
+    File databaseFolder = new File(hive.warehouseDir(), databaseName);
     assertThat(new File(database.getLocationUri()) + "/", is(databaseFolder.toURI().toString()));
   }
 
   @Test
   public void customProperties() {
-    Map<String, String> conf = new HashMap<>();
-    conf.put("my.custom.key", "my.custom.value");
-    HiveConf hiveConf = new HiveMetaStoreJUnitRule("db", conf).conf();
+    HiveConf hiveConf = customPropertiesRule.conf();
     assertThat(hiveConf.get("my.custom.key"), is("my.custom.value"));
   }
 
   @Test(expected = AlreadyExistsException.class)
   public void createExistingDatabase() throws TException {
-    hiveDefaultName.createDatabase(hiveDefaultName.databaseName());
+    defaultDbRule.createDatabase(defaultDbRule.databaseName());
   }
 
   @Test(expected = NullPointerException.class)
   public void createDatabaseNullName() throws TException {
-    hiveDefaultName.createDatabase(null);
+    defaultDbRule.createDatabase(null);
   }
 
   @Test(expected = InvalidObjectException.class)
   public void createDatabaseInvalidName() throws TException {
-    hiveDefaultName.createDatabase("");
+    defaultDbRule.createDatabase("");
   }
 
   @AfterClass
   public static void afterClass() {
-    assertFalse(tempRoot.exists());
+    assertFalse("Found folder at " + tempRoot, tempRoot.exists());
   }
 }