+++ src/test/java/com/hotels/beeju/core/BeejuCoreTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2015-2020 Expedia, Inc.
+ * Copyright (C) 2015-2021 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -29,20 +29,22 @@ import org.apache.derby.jdbc.EmbeddedDriver;
 import org.apache.hadoop.hive.conf.HiveConf;
 import org.apache.hadoop.hive.metastore.HiveMetaStoreClient;
 import org.apache.hadoop.hive.metastore.api.Database;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 
 public class BeejuCoreTest {
-  
-  private String preKey = "my.custom.pre.key"; 
+
+  private String preKey = "my.custom.pre.key";
   private String preValue = "my.custom.pre.value";
-  private String postKey = "my.custom.post.key"; 
-  private String postValue = "my.custom.post.value"; 
+  private String postKey = "my.custom.post.key";
+  private String postValue = "my.custom.post.value";
   private String coreOverrideValue = "user-that-core-will-override";
   private String confOverrideValue = "password-that-will-override-core";
 
   private final BeejuCore defaultCore = new BeejuCore();
   private final BeejuCore dbNameCore = new BeejuCore("test_db");
-  private final BeejuCore dbNameAndMapConfCore = new BeejuCore("test_db_2", createPreConfigurationMap(), createPostConfigurationMap());
+  private final BeejuCore dbNameAndMapConfCore = new BeejuCore("test_db_2", createPreConfigurationMap(),
+      createPostConfigurationMap());
   private final BeejuCore dbNameAndHiveConfCore = new BeejuCore("test_db_2", createPreHiveConf(), createPostHiveConf());
 
   private Map<String, String> createPreConfigurationMap() {
@@ -51,14 +53,14 @@ public class BeejuCoreTest {
     conf.put(HiveConf.ConfVars.METASTORE_CONNECTION_USER_NAME.toString(), coreOverrideValue);
     return conf;
   }
-  
+
   private Map<String, String> createPostConfigurationMap() {
     Map<String, String> conf = new HashMap<>();
     conf.put(postKey, postValue);
     conf.put(HiveConf.ConfVars.METASTOREPWD.toString(), confOverrideValue);
     return conf;
   }
-  
+
   private HiveConf createPreHiveConf() {
     HiveConf conf = new HiveConf();
     conf.clear();
@@ -66,13 +68,21 @@ public class BeejuCoreTest {
     conf.setVar(HiveConf.ConfVars.METASTORE_CONNECTION_USER_NAME, coreOverrideValue);
     return conf;
   }
-  
+
   private HiveConf createPostHiveConf() {
     HiveConf conf = new HiveConf();
     conf.clear();
     conf.set(postKey, postValue);
     conf.setVar(HiveConf.ConfVars.METASTOREPWD, confOverrideValue);
     return conf;
+  }
+
+  @AfterEach
+  public void cleanUp() {
+    defaultCore.cleanUp();
+    dbNameCore.cleanUp();
+    dbNameAndMapConfCore.cleanUp();
+    dbNameAndHiveConfCore.cleanUp();
   }
 
   @Test
@@ -90,12 +100,12 @@ public class BeejuCoreTest {
     assertThat(dbNameAndMapConfCore.databaseName(), is("test_db_2"));
     assertThat(dbNameAndMapConfCore.conf().get(preKey), is(preValue));
     assertThat(dbNameAndMapConfCore.conf().get(postKey), is(postValue));
-    //below still set to BeeJU default as pre-config not overridden
+    // below still set to BeeJU default as pre-config not overridden
     assertThat(dbNameAndMapConfCore.conf().getVar(HiveConf.ConfVars.METASTORE_CONNECTION_USER_NAME), is("db_user"));
-    //below overridden by post-config
+    // below overridden by post-config
     assertThat(dbNameAndMapConfCore.conf().getVar(HiveConf.ConfVars.METASTOREPWD), is(confOverrideValue));
   }
-  
+
   @Test
   public void initialisedDbNameAndHiveConfConstructor() {
     assertThat(dbNameAndHiveConfCore.databaseName(), is("test_db_2"));
@@ -108,14 +118,15 @@ public class BeejuCoreTest {
   }
 
   @Test
-  public void createDirectory(){
-    assertThat(defaultCore.conf().getVar(HiveConf.ConfVars.METASTOREWAREHOUSE), is(defaultCore.tempDir().toString()));
+  public void createDirectory() {
+    assertThat(defaultCore.conf().getVar(HiveConf.ConfVars.METASTOREWAREHOUSE), is(defaultCore.warehouseDir().toString()));
   }
 
   @Test
   public void deleteDirectory() throws IOException {
     BeejuCore testCore = new BeejuCore();
     testCore.cleanUp();
+    assertFalse(Files.exists(testCore.warehouseDir()));
     assertFalse(Files.exists(testCore.tempDir()));
   }
 
@@ -162,7 +173,7 @@ public class BeejuCoreTest {
 
     assertThat(db, is(notNullValue()));
     assertThat(db.getName(), is(databaseName.toLowerCase()));
-    assertThat(db.getLocationUri(), is(String.format("file:%s/%s", defaultCore.tempDir(), databaseName)));
+    assertThat(db.getLocationUri(), is(String.format("file:%s/%s", defaultCore.warehouseDir(), databaseName)));
   }
-
+  
 }