+++ src/test/java/com/hotels/beeju/core/ThriftHiveMetaStoreCoreTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2015-2020 Expedia, Inc.
+ * Copyright (C) 2015-2021 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,12 +23,18 @@ import java.util.List;
 
 import org.apache.hadoop.hive.conf.HiveConf;
 import org.apache.hadoop.hive.metastore.HiveMetaStoreClient;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 
 public class ThriftHiveMetaStoreCoreTest {
 
   private final BeejuCore core = new BeejuCore();
   private final ThriftHiveMetaStoreCore thriftHiveMetaStoreCore = new ThriftHiveMetaStoreCore(core);
+  
+  @AfterEach
+  public void cleanUp() {
+    core.cleanUp();
+  }
 
   @Test
   public void before() throws Exception {