+++ src/test/java/com/hotels/beeju/extensions/HiveServer2JUnitExtensionTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2015-2020 Expedia, Inc.
+ * Copyright (C) 2015-2021 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -30,23 +30,34 @@ public class HiveServer2JUnitExtensionTest {
   private static final String DATABASE = "my_test_db";
 
   @RegisterExtension
-  HiveServer2JUnitExtension server = new HiveServer2JUnitExtension(DATABASE);
+  HiveServer2JUnitExtension defaultDbExtension = new HiveServer2JUnitExtension();
+  
+  @RegisterExtension
+  HiveServer2JUnitExtension customDbExtension = new HiveServer2JUnitExtension(DATABASE);
 
-  @Test
-  public void defaultDatabaseName() {
-    String defaultDbName = new HiveServer2JUnitExtension().databaseName();
-    assertThat(defaultDbName, is("test_database"));
+  @RegisterExtension
+  HiveServer2JUnitExtension customPropertiesExtension = new HiveServer2JUnitExtension("custom_props_database",
+      customConfProperties());
+
+  private Map<String, String> customConfProperties() {
+    Map<String, String> conf = Collections.singletonMap("my.custom.key", "my.custom.value");
+    return conf;
   }
 
   @Test
   public void databaseName() {
-    assertThat(server.databaseName(), is(DATABASE));
+    assertThat(customDbExtension.databaseName(), is(DATABASE));
   }
 
   @Test
-  public void customProperties() {
-    Map<String, String> conf = Collections.singletonMap("my.custom.key", "my.custom.value");
-    HiveConf hiveConf = new HiveServer2JUnitExtension(DATABASE, conf).conf();
+  public void defaultDatabaseName() throws Exception {
+    String defaultDbName = defaultDbExtension.databaseName();
+    assertThat(defaultDbName, is("test_database"));
+  }
+
+  @Test
+  public void customProperties() throws Exception {
+    HiveConf hiveConf = customPropertiesExtension.conf();
     assertThat(hiveConf.get("my.custom.key"), is("my.custom.value"));
   }
 }