+++ CHANGELOG.md
@@ -1,5 +1,22 @@
-## [3.4.0] - TBD
+## [4.0.0] - TBD
+### Fixed
+- Intermediate temporary folders now cleaned up as part of test lifecycle.
+- `HiveMetaStoreCore` performs a null check before attempting to close its metastore client.
+
 ### Changed
+- Hive version updated to `2.3.8` (was `2.3.7`).
+- `hive-exec` dependency changed to use `core` classifier to allow source code retrieval for easier debugging.
+- Hive's `METASTORE_CONNECTION_POOLING_TYPE` is now set to `"NONE"`.
+- Hive's `HIVESTATSAUTOGATHER` is now set to `false`.
+- Hive's `HIVE_SERVER2_LOGGING_OPERATION_ENABLED` is now set to `false`.
+- Various Derby, scratch and similar folders now configured to use a base temporary folder.
+- JUnit4 Rules now extend `TestWatcher` instead of `ExternalResource` to allow cleanup of temporary folders on test failures.
+- `BeejuCore` `tempDir()` returns base temporary test folder instead of temporary Hive warehouse dir.
+- `BeejuCore` returns temporary Hive warehouse dir via `warehouseDir()`.
+- JUnit5 Extensions `getTempDirectory()` returns base temporary test folder instead of temporary Hive warehouse dir.
+- JUnit5 Extensions return temporary Hive warehouse dir via `getWarehouseDirectory()`.
+- JUnit4 Rules `tempDir()` returns base temporary test folder instead of temporary Hive warehouse dir.
+- JUnit4 rules return temporary Hive warehouse dir via `warehouseDir()`.
 - `hotels-oss-parent` version updated to `6.1.0` (was `5.0.0`) to enable building with Java 11.
 
 ## [3.3.0] - 2020-10-23